backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9         
channels>=4.0                
daphne>=4.0                  
redis>=5.0                  
docker>=6.1                  
python-dotenv>=1.0           
gunicorn>=21.2               
django-cors-headers>=4.3     
djangorestframework-simplejwt>=5.3 
drf-yasg>=1.21                     
celery>=5.3
